Genetics and gameloop
Now genetics are stored in the butterfly object (or animal object in the future?).
The old "butterflies[]" variable is removed. It is easier to target butterfly objects directly, and probably best practice.
Genetics now effect blendIn, calculated as a percentage. There is no min or max limit for common genetic types. Is this something bad?
At end of each "generation", butterfly pos gets randomized.