Fixed dynamic scaling.. Maybe
I think I fixed it. More testing and ironing bugs will have to be done. I calculated dpi stuff and changed expected screen-size according to that. I want to test a transition with a device of another dpi, tho. Just to see if it really works. And also, UI needs fixing. And I need to make it so it doesn't try changing the screen every frame.